# api/docs/openapi.json
{
  "components": {
    "schemas": {
      "auth.CreateUserParams": {
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "return_to": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "auth.ForgotPasswordParams": {
        "properties": {
          "email": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "auth.LoginParams": {
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "totp": {
            "type": "integer"
          }
        },
        "type": "object"
      },
      "auth.PasswordResetParams": {
        "properties": {
          "new_password": {
            "type": "string"
          },
          "token": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "auth.RefreshParams": {
        "properties": {
          "refresh_token": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "auth.SessionTokenResponse": {
        "properties": {
          "access_token": {
            "type": "string"
          },
          "expires_in": {
            "type": "integer"
          },
          "refresh_token": {
            "type": "string"
          },
          "token_type": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "auth.VerifyEmailParams": {
        "properties": {
          "token": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "auth.VerifyEmailResponse": {
        "properties": {
          "continue_url": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "description": "Type \"Bearer\" followed by a space and JWT token.",
        "in": "header",
        "name": "Authorization",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "api for auth.hisbaan.com",
    "title": "auth",
    "version": "0.0.1"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/auth/forgot-password": {
      "post": {
        "description": "Sends a password reset email to the user",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/auth.ForgotPasswordParams"
              }
            }
          },
          "description": "Email address",
          "required": true,
          "x-originalParamName": "request"
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "summary": "Request password reset",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "description": "Authenticates user and returns access and refresh tokens",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/auth.LoginParams"
              }
            }
          },
          "description": "Login credentials",
          "required": true,
          "x-originalParamName": "request"
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/auth.SessionTokenResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "summary": "Login user",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/auth/password-reset": {
      "post": {
        "description": "Resets user password using a valid reset token",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/auth.PasswordResetParams"
              }
            }
          },
          "description": "Password reset details",
          "required": true,
          "x-originalParamName": "request"
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "summary": "Reset password",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "description": "Issues a new access token using a valid refresh token",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/auth.RefreshParams"
              }
            }
          },
          "description": "Refresh token",
          "required": true,
          "x-originalParamName": "request"
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/auth.SessionTokenResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "summary": "Refresh access token",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/register": {
      "post": {
        "description": "Creates a new user account and sends verification email",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/auth.CreateUserParams"
              }
            }
          },
          "description": "User registration details",
          "required": true,
          "x-originalParamName": "request"
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request"
          },
          "409": {
            "description": "Conflict"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "summary": "Register a new user",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/verify-email": {
      "post": {
        "description": "Verifies user email using a valid verification token",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/auth.VerifyEmailParams"
              }
            }
          },
          "description": "Email verification token",
          "required": true,
          "x-originalParamName": "request"
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/auth.VerifyEmailResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "summary": "Verify email",
        "tags": [
          "auth"
        ]
      }
    },
...
  }
}